+++ tests/global_configs/globalTeardown.js
@@ -1,0 +1,8 @@
+const opener = require('opener')
+const fs = require('fs-extra')
+module.exports = async function() {
+  // Clean up the test by deleting the app folder
+  fs.removeSync(`${global.appConfig.appPath}`)
+  console.log('Finish up tests')
+  opener('./tests_report/report.html')
+}